[test] Stop opening browser by default in local Dev Overlay Storybook
This doesn't properly work with if a different Chrome release channel is used.
It's also annoying if you restart it and have a tab already open.
We don't open a browser in Next.js for those reasons.

Storybook includes a link to the local version in the terminal so you can always click that.